Move SSW files to an ssw/ subdirectory
