+++ ext/ssw/ssw.c
