+++ ext/ssw/ssw_cpp.cpp
