+++ ext/ssw/ssw_cpp.h
